Introduce dace_common module to share methods between iterator and fieldview backends
