+++ src/gt4py/next/program_processors/runners/dace_fieldview/__init__.py
@@ -7,7 +7,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 
-from gt4py.next.program_processors.runners.dace_fieldview.gtir_dace_backend import get_sdfg_args
+from gt4py.next.program_processors.runners.dace_common.dace_backend import get_sdfg_args
 from gt4py.next.program_processors.runners.dace_fieldview.gtir_to_sdfg import build_sdfg_from_gtir
