+++ src/gt4py/next/program_processors/runners/dace_fieldview/gtir_to_tasklet.py
@@ -19,6 +19,7 @@ from gt4py.next import common as gtx_common
 from gt4py.next.iterator import ir as gtir
 from gt4py.next.iterator.ir_utils import common_pattern_matcher as cpm
 from gt4py.next.iterator.type_system import type_specifications as gtir_ts
+from gt4py.next.program_processors.runners.dace_common import utility as dace_common_util
 from gt4py.next.program_processors.runners.dace_fieldview import (
     gtir_python_codegen,
     gtir_to_sdfg,
@@ -195,7 +196,7 @@ class LambdaToTasklet(eve.NodeVisitor):
     ) -> ValueExpr:
         temp_name = self.sdfg.temp_data_name()
         self.sdfg.add_scalar(temp_name, dtype, transient=True)
-        data_type = dace_fieldview_util.as_scalar_type(str(dtype.as_numpy_dtype()))
+        data_type = dace_common_util.as_scalar_type(str(dtype.as_numpy_dtype()))
         temp_node = self.state.add_access(temp_name)
         self._add_edge(
             src_node,
@@ -323,7 +324,7 @@ class LambdaToTasklet(eve.NodeVisitor):
         assert all(isinstance(index, SymbolExpr) for index in it.indices.values())
 
         field_desc = it.field.desc(self.sdfg)
-        connectivity = dace_fieldview_util.connectivity_identifier(offset)
+        connectivity = dace_common_util.connectivity_identifier(offset)
         # initially, the storage for the connectivty tables is created as transient;
         # when the tables are used, the storage is changed to non-transient,
         # as the corresponding arrays are supposed to be allocated by the SDFG caller
@@ -692,7 +693,7 @@ class LambdaToTasklet(eve.NodeVisitor):
             # initially, the storage for the connectivity tables is created as transient;
             # when the tables are used, the storage is changed to non-transient,
             # so the corresponding arrays are supposed to be allocated by the SDFG caller
-            offset_table = dace_fieldview_util.connectivity_identifier(offset)
+            offset_table = dace_common_util.connectivity_identifier(offset)
             self.sdfg.arrays[offset_table].transient = False
             offset_table_node = self.state.add_access(offset_table)
