+++ src/gt4py/next/program_processors/runners/dace_iterator/itir_to_tasklet.py
@@ -22,13 +22,12 @@ from gt4py.next.common import _DEFAULT_SKIP_VALUE as neighbor_skip_value, Connec
 from gt4py.next.iterator import ir as itir
 from gt4py.next.iterator.ir import FunCall, Lambda
 from gt4py.next.iterator.type_system import type_specifications as it_ts
+from gt4py.next.program_processors.runners.dace_common import utility as dace_common_util
 from gt4py.next.type_system import type_specifications as ts
 
 from .utility import (
     add_mapped_nested_sdfg,
     as_dace_type,
-    connectivity_identifier,
-    dace_debuginfo,
     flatten_list,
     get_used_connectivities,
     map_nested_sdfg_symbols,
@@ -233,7 +232,9 @@ def _visit_lift_in_neighbors_reduction(
             input_nodes[x] = y.value
 
     neighbor_tables = get_used_connectivities(node.args[0], transformer.offset_provider)
-    connectivity_names = [connectivity_identifier(offset) for offset in neighbor_tables.keys()]
+    connectivity_names = [
+        dace_common_util.connectivity_identifier(offset) for offset in neighbor_tables.keys()
+    ]
 
     parent_sdfg = transformer.context.body
     parent_state = transformer.context.state
@@ -326,7 +327,7 @@ def builtin_neighbors(
     sdfg: dace.SDFG = transformer.context.body
     state: dace.SDFGState = transformer.context.state
 
-    di = dace_debuginfo(node, sdfg.debuginfo)
+    di = dace_common_util.debug_info(node, default=sdfg.debuginfo)
     offset_literal, data = node_args
     assert isinstance(offset_literal, itir.OffsetLiteral)
     offset_dim = offset_literal.value
@@ -377,7 +378,7 @@ def builtin_neighbors(
         debuginfo=di,
     )
 
-    table_name = connectivity_identifier(offset_dim)
+    table_name = dace_common_util.connectivity_identifier(offset_dim)
     shift_tasklet = state.add_tasklet(
         "shift",
         code=f"__result = __table[__idx, {neighbor_map_index}]",
@@ -516,7 +517,7 @@ def builtin_neighbors(
 def builtin_can_deref(
     transformer: "PythonTaskletCodegen", node: itir.Expr, node_args: list[itir.Expr]
 ) -> list[ValueExpr]:
-    di = dace_debuginfo(node, transformer.context.body.debuginfo)
+    di = dace_common_util.debug_info(node, default=transformer.context.body.debuginfo)
     # first visit shift, to get set of indices for deref
     can_deref_callable = node_args[0]
     assert isinstance(can_deref_callable, itir.FunCall)
@@ -679,7 +680,7 @@ def builtin_if(
 def builtin_list_get(
     transformer: "PythonTaskletCodegen", node: itir.Expr, node_args: list[itir.Expr]
 ) -> list[ValueExpr]:
-    di = dace_debuginfo(node, transformer.context.body.debuginfo)
+    di = dace_common_util.debug_info(node, default=transformer.context.body.debuginfo)
     args = list(itertools.chain(*transformer.visit(node_args)))
     assert len(args) == 2
     # index node
@@ -705,7 +706,7 @@ def builtin_list_get(
 def builtin_cast(
     transformer: "PythonTaskletCodegen", node: itir.Expr, node_args: list[itir.Expr]
 ) -> list[ValueExpr]:
-    di = dace_debuginfo(node, transformer.context.body.debuginfo)
+    di = dace_common_util.debug_info(node, default=transformer.context.body.debuginfo)
     args = transformer.visit(node_args[0])
     internals = [f"{arg.value.data}_v" for arg in args]
     target_type = node_args[1]
@@ -720,7 +721,7 @@ def builtin_cast(
 def builtin_make_const_list(
     transformer: "PythonTaskletCodegen", node: itir.Expr, node_args: list[itir.Expr]
 ) -> list[ValueExpr]:
-    di = dace_debuginfo(node, transformer.context.body.debuginfo)
+    di = dace_common_util.debug_info(node, default=transformer.context.body.debuginfo)
     args = [transformer.visit(arg)[0] for arg in node_args]
     assert all(isinstance(x, (SymbolExpr, ValueExpr)) for x in args)
     args_dtype = [x.dtype for x in args]
@@ -914,11 +915,15 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
         neighbor_tables = (
             get_used_connectivities(node, self.offset_provider) if use_neighbor_tables else {}
         )
-        connectivity_names = [connectivity_identifier(offset) for offset in neighbor_tables.keys()]
+        connectivity_names = [
+            dace_common_util.connectivity_identifier(offset) for offset in neighbor_tables.keys()
+        ]
 
         # Create the SDFG for the lambda's body
         lambda_sdfg = dace.SDFG(func_name)
-        lambda_sdfg.debuginfo = dace_debuginfo(node, self.context.body.debuginfo)
+        lambda_sdfg.debuginfo = dace_common_util.debug_info(
+            node, default=self.context.body.debuginfo
+        )
         lambda_state = lambda_sdfg.add_state(f"{func_name}_body", is_start_block=True)
 
         lambda_symbols_pass = GatherLambdaSymbolsPass(
@@ -1064,7 +1069,7 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
 
         neighbor_tables = get_used_connectivities(node.fun, self.offset_provider)
         for offset in neighbor_tables.keys():
-            var = connectivity_identifier(offset)
+            var = dace_common_util.connectivity_identifier(offset)
             nsdfg_inputs[var] = dace.Memlet.from_array(var, self.context.body.arrays[var])
 
         symbol_mapping = map_nested_sdfg_symbols(self.context.body, func_context.body, nsdfg_inputs)
@@ -1075,7 +1080,7 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
             inputs=set(nsdfg_inputs.keys()),
             outputs=set(r.value.data for r in results),
             symbol_mapping=symbol_mapping,
-            debuginfo=dace_debuginfo(node, func_context.body.debuginfo),
+            debuginfo=dace_common_util.debug_info(node, default=func_context.body.debuginfo),
         )
 
         for name, value in func_inputs:
@@ -1093,7 +1098,7 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
                     idx_memlet = nsdfg_inputs[var]
                     self.context.state.add_edge(store, None, nsdfg_node, var, idx_memlet)
         for offset in neighbor_tables.keys():
-            var = connectivity_identifier(offset)
+            var = dace_common_util.connectivity_identifier(offset)
             memlet = nsdfg_inputs[var]
             access = self.context.state.add_access(var, debuginfo=nsdfg_node.debuginfo)
             self.context.state.add_edge(access, None, nsdfg_node, var, memlet)
@@ -1110,7 +1115,7 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
         return result_exprs
 
     def _visit_deref(self, node: itir.FunCall) -> list[ValueExpr]:
-        di = dace_debuginfo(node, self.context.body.debuginfo)
+        di = dace_common_util.debug_info(node, default=self.context.body.debuginfo)
         iterator = self.visit(node.args[0])
         if not isinstance(iterator, IteratorExpr):
             # already a list of ValueExpr
@@ -1187,7 +1192,7 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
         )
 
     def _visit_shift(self, node: itir.FunCall) -> IteratorExpr | list[ValueExpr]:
-        di = dace_debuginfo(node, self.context.body.debuginfo)
+        di = dace_common_util.debug_info(node, default=self.context.body.debuginfo)
         shift = node.fun
         assert isinstance(shift, itir.FunCall)
         tail, rest = self._split_shift_args(shift.args)
@@ -1211,7 +1216,7 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
         if isinstance(self.offset_provider[offset_dim], Connectivity):
             offset_provider = self.offset_provider[offset_dim]
             connectivity = self.context.state.add_access(
-                connectivity_identifier(offset_dim), debuginfo=di
+                dace_common_util.connectivity_identifier(offset_dim), debuginfo=di
             )
 
             shifted_dim = offset_provider.origin_axis.value
@@ -1243,7 +1248,7 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
         return IteratorExpr(iterator.field, shifted_index, iterator.dtype, iterator.dimensions)
 
     def visit_OffsetLiteral(self, node: itir.OffsetLiteral) -> list[ValueExpr]:
-        di = dace_debuginfo(node, self.context.body.debuginfo)
+        di = dace_common_util.debug_info(node, default=self.context.body.debuginfo)
         offset = node.value
         assert isinstance(offset, int)
         offset_var = unique_var_name()
@@ -1258,7 +1263,7 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
         return [ValueExpr(offset_node, self.context.body.arrays[offset_var].dtype)]
 
     def _visit_reduce(self, node: itir.FunCall):
-        di = dace_debuginfo(node, self.context.body.debuginfo)
+        di = dace_common_util.debug_info(node, default=self.context.body.debuginfo)
         reduce_dtype = itir_type_as_dace_type(node.type)  # type: ignore[arg-type]  # ensure by type inference
 
         if len(node.args) == 1:
@@ -1446,7 +1451,7 @@ class PythonTaskletCodegen(gt4py.eve.codegen.TemplatedGenerator):
             expr,
             type_,
             "numeric",
-            dace_debuginfo=dace_debuginfo(node, self.context.body.debuginfo),
+            dace_debuginfo=dace_common_util.debug_info(node, default=self.context.body.debuginfo),
         )
 
     def _visit_general_builtin(self, node: itir.FunCall) -> list[ValueExpr]:
@@ -1509,7 +1514,7 @@ def closure_to_tasklet_sdfg(
     use_field_canonical_representation: bool,
 ) -> tuple[Context, Sequence[ValueExpr]]:
     body = dace.SDFG("tasklet_toplevel")
-    body.debuginfo = dace_debuginfo(node)
+    body.debuginfo = dace_common_util.debug_info(node)
     state = body.add_state("tasklet_toplevel_entry", True)
     symbol_map: dict[str, TaskletExpr] = {}
